refactor: updateStyles into own file
